fixed png rendering (not smooth)

+++ Mapsui.Rendering.Skia/MapRenderer.cs
@@ -108,8 +108,8 @@ namespace Mapsui.Rendering.Skia
                 switch (streamFormat)
                 {
                     case EStreamFormat.Png:
-                        var width = (int)viewport.Width;
-                        var height = (int)viewport.Height;
+                        var width = viewport.Width;
+                        var height = viewport.Height;
                         var imageInfo = new SKImageInfo((int)Math.Round(width * pixelDensity), (int)Math.Round(height * pixelDensity),
                             SKImageInfo.PlatformColorType, SKAlphaType.Unpremul);
